refactor: factory(생성 클래스)는 factory 패키지 하위로 모아라
