+++ src/test/java/com/camping/admin/helper/factory/TokenFactory.java
@@ -1,4 +1,4 @@
-package com.camping.admin.helper;
+package com.camping.admin.helper.factory;
 
 import java.util.Map;
