+++ src/test/java/com/camping/admin/steps/Hooks.java
@@ -4,8 +4,8 @@ import com.camping.admin.helper.CommonContext;
 import io.cucumber.java.Before;
 import io.cucumber.java.BeforeAll;
 
-import static com.camping.admin.helper.RequestSpecFactory.getRequestSpec;
-import static com.camping.admin.helper.TokenFactory.getAdminToken;
+import static com.camping.admin.helper.factory.RequestSpecFactory.getRequestSpec;
+import static com.camping.admin.helper.factory.TokenFactory.getAdminToken;
 
 /**
  * 비즈니스 로직에 영향을 주지 않는 공통 세팅 (예. 토큰 세팅)
